refactor: organize receipt directory buy domain
