--- src/Receipt/DetailGroup/Donation.php
@@ -1,19 +1,0 @@
-<?php
-namespace Give\Receipt\DetailGroup;
-
-use Give\Receipt\Detail\Donation\Amount;
-use Give\Receipt\Detail\Donation\PaymentGateway;
-use Give\Receipt\Detail\Donation\Status;
-use Give\Receipt\Detail\Donation\TotalAmount;
-use Give\Receipt\DetailGroup;
-
-class Donation extends DetailGroup {
-	public $groupId = 'donationDetails';
-
-	protected $detailsList = [
-		PaymentGateway::class,
-		Status::class,
-		Amount::class,
-		TotalAmount::class,
-	];
-}
