+++ src/Receipt/DonationDetailsGroup/Details/PaymentGateway.php
@@ -1,5 +1,5 @@
 <?php
-namespace Give\Receipt\Detail\Donation;
+namespace Give\Receipt\DonationDetailsGroup\Details;
 
 use Give\Receipt\Detail;
 use function give_get_payment_meta as getDonationMetaData;
